# backend/src/docs/swagger/index.yaml
openapi: 3.0.0
info:
  title: Social Developer API
  version: 1.0.0
  description: |
    API documentation for the Social Developer platform.
    This API provides endpoints for managing social media automation tasks, payments, and user accounts.
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.socialdeveloper.com/api
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        service:
          type: string
        platform:
          type: string
          enum: [instagram, youtube]
        targetUrl:
          type: string
        quantity:
          type: integer
        startCount:
          type: integer
        remainingCount:
          type: integer
        status:
          type: string
          enum: [pending, processing, completed, failed]
        speed:
          type: string
          enum: [normal, fast, express]
        amount:
          type: number
          format: float
...
paths:
...
  /orders:
    get:
      tags:
        - Orders
      summary: Get user's orders
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      totalPages:
                        type: integer

    post:
      tags:
        - Orders
      summary: Create a new order
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - platform
                - service
                - targetUrl
                - quantity
              properties:
                platform:
                  type: string
                  enum: [instagram, youtube]
                service:
                  type: string
                targetUrl:
                  type: string
                quantity:
                  type: integer
                speed:
                  type: string
                  enum: [normal, fast, express]
                  default: normal
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
...